# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "charpente API",
    "description": "Backend API. Errors are RFC 9457 problem+json with a stable `code`.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/account": {
      "delete": {
        "tags": [
          "account"
        ],
        "summary": "Erase the account (GDPR art. 17). Deletes the user row; uploads cascade.",
        "operationId": "delete_account",
        "responses": {
          "204": {
            "description": "Account erased"
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/account/export": {
      "get": {
        "tags": [
          "account"
        ],
        "summary": "Export the account's data as JSON (GDPR art. 20).",
        "operationId": "export_account",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountExport"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/admin/audit-log": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "The audit trail of role-guarded actions, newest first.",
        "operationId": "list_audit_log",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 25).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditLogEntryDto"
                }
              }
            }
          },
          "403": {
            "description": "Missing permission"
          }
        }
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List all users (admin only), newest first, cursor-paginated.",
        "operationId": "list_users",
        "parameters": [
          {
            "name": "cursor",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Page size, 1..=100 (default 25).",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          },
          "403": {
            "description": "Not an admin"
          }
        }
      }
    },
    "/api/v1/admin/users/{id}/role": {
      "patch": {
        "tags": [
          "admin"
        ],
        "summary": "Change a user's role (admin only). Recorded in the audit log.",
        "operationId": "set_user_role",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "User id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetRoleRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "403": {
            "description": "Not an admin"
          },
          "404": {
            "description": "No such user"
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Log in with email + password.",
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "tags": [
          "auth"
        ],
        "summary": "The currently authenticated user.",
        "operationId": "me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "401": {
            "description": "Not authenticated"
          }
        }
      }
    },
    "/api/v1/auth/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Create an account and open a session.",
        "operationId": "register",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Account created and logged in",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "409": {
            "description": "Email already registered"
          },
          "422": {
            "description": "Validation failed"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountExport": {
        "type": "object",
        "required": [
          "user",
          "uploads"
        ],
        "properties": {
          "uploads": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UploadDto"
            }
          },
          "user": {
            "$ref": "#/components/schemas/UserDto"
          }
        }
      },
...
      "LoginRequest": {
        "type": "object",
        "required": [
          "email",
          "password"
        ],
        "properties": {
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "Page_AuditLogEntryDto": {
        "type": "object",
        "required": [
          "items",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "action",
                "detail",
                "created_at"
              ],
              "properties": {
                "action": {
                  "type": "string"
                },
                "actor_id": {
                  "type": [
                    "string",
                    "null"
                  ],
                  "format": "uuid"
                },
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "detail": {},
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "subject": {
                  "type": [
                    "string",
                    "null"
                  ]
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Pass as `cursor` to fetch the next page; absent on the last page."
          }
        }
      },
...
      "Page_UserDto": {
        "type": "object",
        "required": [
          "items",
          "has_more"
        ],
        "properties": {
          "has_more": {
            "type": "boolean"
          },
          "items": {
            "type": "array",
            "items": {
              "type": "object",
              "required": [
                "id",
                "email",
                "display_name",
                "role",
                "created_at"
              ],
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time"
                },
                "display_name": {
                  "type": "string"
                },
                "email": {
                  "type": "string"
                },
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "role": {
                  "$ref": "#/components/schemas/RoleDto"
                }
              }
            }
          },
          "next_cursor": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Pass as `cursor` to fetch the next page; absent on the last page."
          }
        }
      },
...
      "RegisterRequest": {
        "type": "object",
        "required": [
          "email",
          "password",
          "display_name"
        ],
        "properties": {
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      },
      "RoleDto": {
        "type": "string",
        "enum": [
          "admin",
          "user"
        ]
      },
      "SetRoleRequest": {
        "type": "object",
        "required": [
          "role"
        ],
        "properties": {
          "role": {
            "$ref": "#/components/schemas/RoleDto"
          }
        }
      },
      "UploadDto": {
        "type": "object",
        "required": [
          "id",
          "filename",
          "content_type",
          "size_bytes",
          "state",
          "created_at"
        ],
        "properties": {
          "content_type": {
            "type": "string"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "filename": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "size_bytes": {
            "type": "integer",
            "format": "int64"
          },
          "state": {
            "$ref": "#/components/schemas/UploadState"
          }
        }
      },
      "UploadState": {
        "type": "string",
        "enum": [
          "pending",
          "complete"
        ]
      },
      "UserDto": {
        "type": "object",
        "required": [
          "id",
          "email",
          "display_name",
          "role",
          "created_at"
        ],
        "properties": {
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "display_name": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "$ref": "#/components/schemas/RoleDto"
          }
        }
      }
    }
  },
...
}